Ch7 completed refactor to use f64 to avoid acne effect, some more optimization to reduce rendering time

+++ src/camera.rs
@@ -1,6 +1,6 @@
 use crate::canvas::Canvas;
-use crate::matrix::{Mat4, mat4::identity};
-use crate::ray::Ray; 
+use crate::matrix::{mat4::identity, Mat4};
+use crate::ray::Ray;
 use crate::tuple::Point;
 use crate::world::World;
 
@@ -8,34 +8,34 @@ use crate::world::World;
 pub struct Camera {
     pub hsize: usize,
     pub vsize: usize,
-    pub fov: f32,
-    pub half_width: f32,
-    pub half_height: f32,
-    pub pixel_size: f32,
+    pub fov: f64,
+    pub half_width: f64,
+    pub half_height: f64,
+    pub pixel_size: f64,
     transform: Mat4,
     inverse_transform: Mat4,
 }
 
 impl Camera {
-    pub fn new(hsize: usize, vsize: usize, fov: f32) -> Self {
+    pub fn new(hsize: usize, vsize: usize, fov: f64) -> Self {
         let half_view = (fov / 2.).tan();
-        let aspect_ratio = hsize as f32 / vsize as f32;
+        let aspect_ratio = hsize as f64 / vsize as f64;
         let (half_width, half_height) = if aspect_ratio >= 1. {
             (half_view, half_view / aspect_ratio)
         } else {
             (half_view * aspect_ratio, half_view)
         };
-        let pixel_size = half_width * 2. / hsize as f32;
-        Self { 
-            hsize, 
-            vsize, 
+        let pixel_size = half_width * 2. / hsize as f64;
+        Self {
+            hsize,
+            vsize,
             fov,
             half_width,
             half_height,
             pixel_size,
             transform: identity(),
             inverse_transform: identity().inv(),
-        } 
+        }
     }
 
     pub fn transform(&self) -> Mat4 {
@@ -51,16 +51,16 @@ impl Camera {
     /// pass to the given [x, y] pixel on the canvas
     pub fn ray_for_pixel(&self, x: usize, y: usize) -> Ray {
         // the offset from the edge of the canvas to the pixel's center
-        let x_offset = (x as f32 + 0.5) * self.pixel_size;
-        let y_offset = (y as f32 + 0.5) * self.pixel_size;
-        
+        let x_offset = (x as f64 + 0.5) * self.pixel_size;
+        let y_offset = (y as f64 + 0.5) * self.pixel_size;
+
         // the untransformed coordinates of the pixel in world space.
         // (remember that the camera looks toward -z, so +x is to the *left*.)
         let x_world = self.half_width - x_offset;
         let y_world = self.half_height - y_offset;
 
         // using the camera matrix, transform the canvas point and the origin,
-        // and then compute the ray's direction vector. 
+        // and then compute the ray's direction vector.
         // (remember that the canvas is at z=-1)
         let pixel = self.inverse_transform * Point::new(x_world, y_world, -1.0);
         let origin = self.inverse_transform * Point::new(0.0, 0.0, 0.0);
@@ -69,7 +69,7 @@ impl Camera {
     }
 
     /// This function cast a ray through each of world canvas pixels
-    pub fn render(&self, w: World) -> Canvas{
+    pub fn render(&self, w: World) -> Canvas {
         let mut image = Canvas::new(self.hsize, self.vsize);
         for y in 0..self.vsize {
             for x in 0..self.hsize {
@@ -80,4 +80,3 @@ impl Camera {
         image
     }
 }
-
