+++ src/comps.rs
@@ -2,11 +2,11 @@ use crate::shape::Shape;
 use crate::tuple::*;
 
 pub struct Comps {
-    pub t: f32,
+    pub t: f64,
     pub object: Box<dyn Shape>,
     pub point: Point,
     pub eyev: Vector,
-    pub normalv: Vector, 
+    pub normalv: Vector,
     pub inside: bool,
     pub over_point: Point,
 }
